# bff/src/widgets/dto/widget.dto.ts
import { ApiProperty, ApiSchema } from '@nestjs/swagger'

export const WIDGET_STATUSES = ['active', 'inactive', 'archived'] as const
export type WidgetStatus = (typeof WIDGET_STATUSES)[number]

export const WIDGET_EXAMPLE = {
  id: '4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f',
  subject: 'user-123',
  name: 'Intake form',
  description: 'Widget used for intake workflow testing.',
  status: 'active',
  additionalData: {
    source: 'local-dev',
  },
  createdAt: '2026-05-13T18:00:00Z',
  updatedAt: '2026-05-13T18:00:00Z',
}

@ApiSchema({
  name: 'Widget',
  description: 'Widget resource owned by a subject.',
})
export class WidgetDto {
  @ApiProperty({
    format: 'uuid',
    example: WIDGET_EXAMPLE.id,
    readOnly: true,
  })
  id: string

  @ApiProperty({
    description: 'Owner subject from the provider SDX API.',
    example: WIDGET_EXAMPLE.subject,
    readOnly: true,
  })
  subject: string

  @ApiProperty({
    minLength: 1,
    maxLength: 200,
    example: WIDGET_EXAMPLE.name,
  })
  name: string

  @ApiProperty({
    type: 'string',
    nullable: true,
    maxLength: 1000,
    example: WIDGET_EXAMPLE.description,
  })
  description: string | null

  @ApiProperty({
    enum: WIDGET_STATUSES,
    enumName: 'WidgetStatus',
    example: WIDGET_EXAMPLE.status,
  })
  status: WidgetStatus

  @ApiProperty({
    type: 'object',
    additionalProperties: true,
    example: WIDGET_EXAMPLE.additionalData,
  })
  additionalData: Record<string, unknown>

  @ApiProperty({
    format: 'date-time',
    example: WIDGET_EXAMPLE.createdAt,
    readOnly: true,
  })
  createdAt: Date

  @ApiProperty({
    format: 'date-time',
    example: WIDGET_EXAMPLE.updatedAt,
    readOnly: true,
  })
  updatedAt: Date
}

@ApiSchema({
  name: 'ErrorResponse',
  description:
    'Standard error response format for unexpected or server-side errors. The details.correlationId value is used for support and log tracing.',
})
export class ErrorResponseDto {
  @ApiProperty({
    title: 'Error',
    description:
      'A short, machine-readable error code or identifier that categorizes the type of error. This field is stable and intended for programmatic handling by clients.',
    example: 'forbidden',
  })
  error: string

  @ApiProperty({
    title: 'Message',
    description:
      'A human-readable summary of the error, suitable for display to end-users or logging.',
    example: 'You are not authorized to access this resource',
  })
  message: string

  @ApiProperty({
    title: 'Details',
    required: false,
    nullable: true,
    type: Object,
    additionalProperties: true,
    description:
      'Optional additional context or structured details about the error. Use sparingly and avoid including sensitive data. For generated error responses, this object includes a `correlationId`. Values generated by this API use the `widget-<uuid>` format; values supplied by upstream callers are preserved as received.',
    example: {
      correlationId: 'widget-2f7b8d43-7b0d-4b5f-8a6c-1a2b3c4d5e6f',
      timestamp: '2026-01-16T19:22:00Z',
    },
  })
  details?: Record<string, unknown> | null
}

@ApiSchema({
  name: 'ProblemDetailErrorItem',
  description: 'Represents a single detailed error within a Problem Details response (RFC 9457).',
})
export class ProblemDetailErrorItemDto {
  @ApiProperty({
    enum: ['body', 'query', 'header', 'path', 'cookie'],
    enumName: 'ProblemDetailErrorLocation',
    example: 'body',
  })
  location: 'body' | 'query' | 'header' | 'path' | 'cookie'

  @ApiProperty({
    title: 'Code',
    description: 'A machine-readable error code that identifies the specific type of error.',
    example: 'INVALID_REQUEST',
  })
  code: string

  @ApiProperty({
    title: 'Message',
    description: 'A short, human-readable summary of the error.',
    example: 'Invalid request body or parameter',
  })
  message: string

  @ApiProperty({
    title: 'Type',
    description: 'A URI reference or tag that identifies the problem type.',
    example: 'tag:validation-error',
  })
  type: string

  @ApiProperty({
    type: 'string',
    title: 'Field',
    description: 'The name of the specific field/property in the request that caused the error.',
    required: false,
    nullable: true,
    example: 'name',
  })
  field?: string | null

  @ApiProperty({
    type: 'string',
    title: 'Detail',
    description: 'A more detailed human-readable explanation of the error.',
    required: false,
    nullable: true,
    example: 'name must be a non-empty string up to 200 characters',
  })
  detail?: string | null

  @ApiProperty({
    type: 'string',
    title: 'Received',
    description: 'The actual value received by the server that caused the error.',
    required: false,
    nullable: true,
    example: '',
  })
  received?: string | null

  @ApiProperty({
    type: 'string',
    title: 'Pointer',
    description:
      'JSON Pointer (RFC 6901) to the exact location of the error within the request body.',
    required: false,
    nullable: true,
    example: '#/name',
  })
  pointer?: string | null

  @ApiProperty({
    title: 'Constraints',
    description: 'Optional object containing validation constraint violations.',
    required: false,
    nullable: true,
    type: Object,
    additionalProperties: true,
    example: {
      minLength: 1,
      maxLength: 200,
    },
  })
  constraints?: Record<string, unknown> | null
}

@ApiSchema({
  name: 'ProblemDetailResponse',
  description: 'Canonical model for problem details as defined by RFC 9457.',
})
export class ProblemDetailResponseDto {
  @ApiProperty({
    title: 'Type',
    description: 'A URI reference that identifies the problem type.',
    example: 'tag:validation-errors',
  })
  type: string

  @ApiProperty({
    title: 'Title',
    description: 'A short, human-readable summary of the problem type.',
    example: 'Bad Request',
  })
  title: string

  @ApiProperty({
    type: 'integer',
    title: 'Status',
    description: 'A number indicating the HTTP status code generated for this occurrence.',
    example: 400,
  })
  status: number

  @ApiProperty({
    type: 'string',
    title: 'Detail',
    description: 'A human-readable explanation specific to this occurrence of the problem.',
    required: false,
    nullable: true,
    example: 'One or more validation errors occurred',
  })
  detail?: string | null

  @ApiProperty({
    type: ProblemDetailErrorItemDto,
    isArray: true,
    minItems: 1,
    title: 'Errors',
    description: 'A list of individual error occurrences with details and locations.',
  })
  errors: ProblemDetailErrorItemDto[]
}